logs-refactor
