+++ duckbot/__main__.py
@@ -24,11 +24,12 @@ from duckbot import DuckBot
 
 
 def run_duckbot(bot: commands.Bot):
+    bot.load_extension(duckbot.logs.__name__)
+
     if "connection-test" in os.getenv("DUCKBOT_ARGS", ""):
         bot.load_extension(duckbot.util.connection_test.__name__)
 
     bot.load_extension(duckbot.health.__name__)
-    bot.load_extension(duckbot.logs.__name__)
 
     bot.load_extension(duckbot.cogs.duck.__name__)
     bot.load_extension(duckbot.cogs.dogs.__name__)
@@ -50,6 +51,5 @@ def run_duckbot(bot: commands.Bot):
 
 
 if __name__ == "__main__":
-    duckbot.logs.define_logs()
     bot = DuckBot()
     run_duckbot(bot)
