+++ duckbot/logs/__init__.py
@@ -1,8 +1,6 @@
-from .define_logs import define_logs
-from .exception_logs import ExceptionLogs
-from .get_logs import GetLogs
+from .logging import Logging
 
 
 def setup(bot):
-    bot.add_cog(ExceptionLogs(bot))
-    bot.add_cog(GetLogs(bot))
+    Logging.define_logs()
+    bot.add_cog(Logging(bot))
