+++ tests/logs/setup_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from duckbot.logs import ExceptionLogs, GetLogs
+from duckbot.logs import Logging
 from duckbot.logs import setup as extension_setup
 from tests.discord_test_ext import assert_cog_added_of_type
 
@@ -8,5 +8,4 @@ from tests.discord_test_ext import assert_cog_added_of_type
 @pytest.mark.asyncio
 async def test_setup(bot_spy):
     extension_setup(bot_spy)
-    assert_cog_added_of_type(bot_spy, GetLogs)
-    assert_cog_added_of_type(bot_spy, ExceptionLogs)
+    assert_cog_added_of_type(bot_spy, Logging)
